chore(type): rename to ir_type
