+++ frontend/ir_type/src/lib.rs
@@ -1,1 +1,1 @@
-pub mod tp;
+pub mod ir_type;
